image_uploader/requirements.txt
click
tqdm
filetype
aiohttp
pytest-mock
pytest-async
pytest-aiohttp
pytest-cov